# swagger.yaml
openapi: 3.0.3
info:
  title: Nexus Project Management API
  description: |
    RESTful API for the Nexus Project Management System.
    
    ## Authentication
    This API uses JWT (JSON Web Token) for authentication. After registering or logging in, 
    you will receive a JWT token that must be included in the `Authorization` header for all 
    protected endpoints.
    
    ## Authorization
    The API implements role-based access control (RBAC) with two roles:
    - **ADMIN**: Full access to all resources
    - **USER**: Limited access to assigned resources only
    
    ## Features
    - User authentication and management
    - Project management with team assignments
    - Task management with status tracking
    - Role-based authorization
    - Pagination and search capabilities
  version: 1.0.0
  contact:
    name: API Support
    email: support@nexus-pm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.nexus-pm.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login or register endpoints.
        
        Example: `Authorization: Bearer <your_jwt_token>`

  # ============================================
  # REUSABLE SCHEMAS
  # ============================================
  schemas:
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - USER
      description: User role for authorization
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Doe
          description: User's full name
        email:
          type: string
          format: email
          example: john.doe@example.com
          description: User's email address (must be unique)
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          example: SecurePass123!
          description: User's password (will be hashed)
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
          description: User's email address
        password:
          type: string
          format: password
          example: SecurePass123!
          description: User's password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          description: JWT token for authentication
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique user identifier
        name:
          type: string
          example: John Doe
          description: User's full name
        email:
          type: string
          format: email
          example: john.doe@example.com
          description: User's email address
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp

    # ========== PROJECT SCHEMAS ==========
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP status code
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: Validation failed
          description: Error message or array of messages
        error:
          type: string
          example: Bad Request
          description: Error type
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        error:
          type: string
          example: Unauthorized
...
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found
        error:
          type: string
          example: Not Found

# ============================================
# API ENDPOINTS
# ============================================
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Create a new user account with email and password.
        
        - Email must be unique
        - Password is automatically hashed using bcrypt
        - Default role is USER
        - Returns user information and JWT token
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              newUser:
                summary: New user registration
                value:
                  name: John Doe
                  email: john.doe@example.com
                  password: SecurePass123!
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Successful registration
                  value:
                    user:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      name: John Doe
                      email: john.doe@example.com
                      role: USER
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-01-15T10:30:00Z"
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validationError:
                  summary: Validation failed
                  value:
                    statusCode: 400
                    message:
                      - "Name must be at least 2 characters long"
                      - "Password must be at least 8 characters long"
                    error: Bad Request
                emailExists:
                  summary: Email already registered
                  value:
                    statusCode: 400
                    message: Email already exists
                    error: Bad Request
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email and password
      description: |
        Authenticate user with email and password.
        
        - Validates user credentials
        - Checks password hash using bcrypt
        - Returns user information and JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              adminLogin:
                summary: Admin login
                value:
                  email: admin@nexus.com
                  password: Admin123!
              userLogin:
                summary: User login
                value:
                  email: john.doe@example.com
                  password: SecurePass123!
      responses:
        '200':
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Successful login
                  value:
                    user:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      name: John Doe
                      email: john.doe@example.com
                      role: USER
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-01-15T10:30:00Z"
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
              examples:
                invalidPassword:
                  summary: Invalid password
                  value:
                    statusCode: 401
                    message: Invalid credentials
                    error: Unauthorized
                userNotFound:
                  summary: User not found
                  value:
                    statusCode: 401
                    message: Invalid credentials
                    error: Unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: |
        Retrieve the authenticated user's profile information.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** All authenticated users
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                userProfile:
                  summary: User profile
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    name: John Doe
                    email: john.doe@example.com
                    role: USER
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-15T10:30:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ========== PROJECT ENDPOINTS ==========
...